swal

+++ src/modules/daybook/views/EntryView.vue
@@ -46,6 +46,7 @@
 	import { defineAsyncComponent } from 'vue';
 	import { mapGetters, mapActions } from 'vuex';
 	import getDayMonthYear from '../helpers/getDayMonthYear';
+	import Swal from 'sweetalert2';
 
 	export default {
 		props: {
@@ -78,6 +79,12 @@
 				this.entry = entry;
 			},
 			async saveEntry() {
+				new Swal({
+					title: 'Espere por favor',
+					allowOutsideClick: false,
+				});
+				Swal.showLoading();
+
 				if (this.entry.id) {
 					// Update
 					await this.updateEntry(this.entry);
@@ -86,10 +93,33 @@
 					const id = await this.createEntry(this.entry);
 					this.$router.push({ name: 'entry', params: { id } });
 				}
+
+				Swal.fire('Guardado', 'Entrada registrada con éxito', 'success');
 			},
 			async onDeleteEntry() {
-				await this.deleteEntry(this.entry.id);
-				this.$router.push({ name: 'no-entry' });
+				const result = await Swal.fire({
+					icon: 'question',
+					title: '¿Está seguro?',
+					text: 'Una vez borrado, no se puede recuperar',
+					showDenyButton: true,
+					confirmButtonText: 'Si, estoy seguro',
+				});
+
+				const { isConfirmed } = result;
+
+				if (isConfirmed) {
+					new Swal({
+						title: 'Espere por favor',
+						allowOutsideClick: false,
+					});
+
+					Swal.showLoading();
+
+					await this.deleteEntry(this.entry.id);
+					this.$router.push({ name: 'no-entry' });
+
+					Swal.fire('Eliminado', '', 'success');
+				}
 			},
 		},
 		computed: {
